requirements.txt
json5>=0.9
PyYAML>=6.0 